% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_ad.tsissm.spec}
\alias{estimate_ad.tsissm.spec}
\alias{estimate_ad}
\title{Estimates an ISSM model given a specification object using maximum likelihood and autodiff}
\usage{
\method{estimate_ad}{tsissm.spec}(
  object,
  solver = "nlminb",
  control = list(trace = 1, eval.max = 300, iter.max = 500),
  use_hessian = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class tsissm.spec.}

\item{solver}{Only \dQuote{nlminb} currently supported.}

\item{control}{Solver control parameters.}

\item{use_hessian}{Whether to include the hessian in the calculation. This is currently
problematic and we suggest not ovveriding the default (FALSE) until further investigation.}

\item{...}{additional parameters passed to the estimation function}
}
\value{
An list of coefficients and other information.
}
\description{
Estimates an ISSM model given a specification object using maximum likelihood and autodiff
}
\details{
This function is not expected to be used by itself but rather as a plugin
to be called from the estimate method of the tsissm package.
}
