% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsissmad-package.R
\docType{package}
\name{tsissmad-package}
\alias{tsissmad}
\alias{tsissmad-package}
\title{tsissmad: ISSM model estimation using automatic differentiation}
\description{
Plugin Estimator for the tsissm package to use autodiff.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

}
\keyword{internal}
